train_stage/requirements.txt
boto3
pandas
numpy
scikit-learn
joblib
flask
tensorflow
pyyaml
minio
